+++ Client/Client/Public/Hawk.h
@@ -146,7 +146,7 @@ public:
 public:
 	virtual _bool Is_AnimationLoop(_uint eAnimId) override;
 
-	virtual _int Take_Damage(float fDamage, CBaseObj* DamageCauser) override;
+	virtual _int Take_Damage(int fDamage, CBaseObj* DamageCauser) override;
 
 	HRESULT CHawk::SetUp_ShaderID();
 private:
